refactor: split useIntegrationActions from useEntryActions
